+++ app/constants.ts
@@ -1,1 +1,2 @@
-export const P2WPKH_ADDRESS = "tb1qe60n447jylrxa96y6pfgy8pq6x9zafu09ky7cq";
+export const nBTC_ADDR = "tb1qe60n447jylrxa96y6pfgy8pq6x9zafu09ky7cq";
+export const MEMPOOL_API = "https://mempool.space/testnet4/api";
